Put fetch weather detail to seperate file and add current tempature in Clock component
