+++ src/components/Clock.tsx
@@ -1,9 +1,13 @@
+import type { WeatherData } from '@/constants/weather';
+import getForecastData from '@/utils/weather';
 import { defineComponent, ref, onMounted, onUnmounted } from 'vue';
 
 export default defineComponent({
   name: 'Clock',
   setup() {
     const time = ref(new Date());
+    const weatherData = ref<WeatherData | null>(null)
+
     let timerInterval: ReturnType<typeof setInterval>;
 
     const weekDays = ['SU', 'MO', 'TU', 'WE', 'TH', 'FR', 'SA'];
@@ -13,7 +17,10 @@ export default defineComponent({
       time.value = new Date();
     };
 
-    onMounted(() => {
+    onMounted(async () => {
+      const hourIndex = Math.trunc(new Date().getHours() / 3)
+      const currentWeatherDetail: WeatherData[] = await getForecastData('Hồ Chí Minh', 0, hourIndex)
+      weatherData.value = currentWeatherDetail[hourIndex]
       updateTime();
       timerInterval = setInterval(updateTime, 400);
     });
@@ -41,13 +48,16 @@ export default defineComponent({
         </div>
         <div class="weekdays">
           {weekDays.map((day, index) => (
-            <span 
-              key={day} 
+            <span
+              key={day}
               class={`weekday ${index === currentDay ? 'active' : ''}`}
             >
               {day}
             </span>
           ))}
+        </div>
+        <div class="tempatuere mt-2">
+          <h4 class="fw-normal">{weatherData.value?.tempature}</h4>
         </div>
       </div>
     );
