+++ src/constants/weather.ts
@@ -1,0 +1,7 @@
+export interface WeatherData {
+    location: string,
+    weather: string,
+    tempature: string,
+    humidity: string,
+    updatedAt: string
+}
